refactor: useUpdateMultipleChoiceAnswer에서 useOptionSelection 분리
- 객관식에서 선택된 문항에 대한 관리 책임은 useOptionSelection에 부여
